star history
Former-commit-id: 7bee856c424bcd4704c4ceb74c759b400a15e2db
Former-commit-id: adbfc8fade4108bef2a32376e416a3c2066f8a23

+++ backend/src/pytorch/InterpolateArchs/util/softsplat_torch.py
@@ -175,7 +175,7 @@ class softsplat_func(torch.autograd.Function):
         # Reshape tenOut back to [N, C, H, W]
         tenOut = tenOut_flat.view(N, H, W, C).permute(0, 3, 1, 2)
 
-        return tenOut
+        return tenOut.to(origdtype)
     # end
